Move SortedResultsClient into a utility module.
